trying a sustainable approach to use duckdb, rsqlite so slow

+++ R/remove.R
@@ -12,8 +12,16 @@
 #' }
 #' }
 censo_borrar_base <- function() {
-  suppressWarnings(censo_desconectar_base())
-  try(unlink(censo_path(), recursive = TRUE))
+  duckdb_version <- utils::packageVersion("duckdb")
+  db_pattern <- paste0("v", gsub("\\.", "", duckdb_version), ".duckdb")
+  
+  existing_files <- list.files(censo_path())
+  
+  if (!any(grepl(db_pattern, existing_files))) {
+    suppressWarnings(censo_desconectar_base())
+    try(unlink(censo_path(), recursive = TRUE))
+  }
+  
   update_censo_pane()
   censo_panel()
 }
